pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datalib"
dynamic = ["version"]
description = "Databricks Data Engineering Library - Reusable data processing components"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.9"
authors = [
    {name = "Data Engineering Team", email = "data-engineering@company.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["databricks", "data-engineering", "etl", "spark"]

dependencies = [
    "pyspark>=3.4.0",
    "pyyaml>=6.0",
    "delta-spark>=2.4.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0",
    "chispa>=0.9.0",  # Spark DataFrame testing
]
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.2",
    "myst-parser>=1.0",
]

[project.urls]
Homepage = "https://github.com/company/databricks-project"
Documentation = "https://github.com/company/databricks-project#readme"
Repository = "https://github.com/company/databricks-project"

[project.scripts]
datalib-version = "datalib.cli:print_version"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["datalib", "datalib.core", "datalib.transformations", "datalib.utils", "datalib.io"]

[tool.setuptools.dynamic]
version = {attr = "datalib._version.__version__"}

[tool.setuptools.package-data]
datalib = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["datalib"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["build", "dist", ".venv"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/datalib"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
